refactor: buttons to use AppButton component
Required reworking the rows and other elements involved to align with
the larger button sizes. Also needed to go to a grid based layout for
inputs & buttons to promote collapsing gracefully on smaller viewports
