+++ frontend/src/components/guiComponents/AppLabel.vue
