+++ frontend/src/components/orgFacilities/ManageOrganization.vue
@@ -1,5 +1,5 @@
 <template>
-  <v-container class="pa-0" fluid>
+  <v-container class="pa-0 text-body-1" fluid>
     <v-row no-gutters>
       <v-col>
         <p>View and update your organization information.</p>
@@ -70,11 +70,11 @@
             <template v-if="editing.email">
               <v-col cols="12" sm="7" xl="8" xxl="9">
                 <v-form v-model="valid.email" @submit.prevent>
-                  <v-row no-gutters justify="space-between">
-                    <v-col>
+                  <v-row no-gutters>
+                    <v-col cols="12" md="6">
                       <v-text-field
                         v-model="workingFields.email"
-                        class="micro"
+                        class="less-jitter"
                         density="compact"
                         variant="underlined"
                         label="Email Address"
@@ -83,30 +83,27 @@
                         hide-details="auto"
                       />
                     </v-col>
-                    <v-col class="ml-2">
-                      <v-btn
-                        size="x-small"
-                        color="#003366"
-                        variant="flat"
-                        prepend-icon="mdi-content-save"
+                    <v-col cols="3" offset-md="1" class="text-no-wrap">
+                      <AppButton
+                        size="small"
                         type="submit"
+                        :display-block="false"
                         :disabled="!valid.email"
                         :loading="isProcessing"
                         @click="() => saveField('email')"
                       >
                         Save
-                      </v-btn>
-                      <v-btn
+                      </AppButton>
+                      <AppButton
                         class="ml-1"
-                        size="x-small"
-                        color="#003366"
-                        variant="outlined"
-                        prepend-icon="mdi-cancel"
+                        size="small"
+                        :primary="false"
+                        :display-block="false"
                         :disabled="isProcessing"
                         @click="editing.email = false"
                       >
                         Cancel
-                      </v-btn>
+                      </AppButton>
                     </v-col>
                   </v-row>
                 </v-form>
@@ -114,21 +111,19 @@
             </template>
             <template v-else>
               <v-col cols="12" sm="7" xl="8" xxl="9">
-                <v-row no-gutters justify="space-between">
-                  <v-col>
+                <v-row no-gutters>
+                  <v-col cols="12" md="6">
                     <p>{{ loadedModel.email }}</p>
                   </v-col>
-                  <v-col class="ml-2">
-                    <v-btn
-                      size="x-small"
-                      color="#003366"
-                      variant="outlined"
-                      prepend-icon="mdi-pencil"
+                  <v-col cols="3" offset-md="1">
+                    <AppButton
+                      size="small"
+                      :display-block="false"
                       :disabled="isProcessing"
                       @click="editing.email = true"
                     >
                       Edit
-                    </v-btn>
+                    </AppButton>
                   </v-col>
                 </v-row>
               </v-col>
@@ -141,11 +136,11 @@
             <template v-if="editing.phone">
               <v-col cols="12" sm="7" xl="8" xxl="9">
                 <v-form v-model="valid.phone" @submit.prevent>
-                  <v-row no-gutters justify="space-between">
-                    <v-col>
+                  <v-row no-gutters>
+                    <v-col cols="12" md="6">
                       <v-text-field
                         v-model="workingFields.phone"
-                        class="micro"
+                        class="less-jitter"
                         density="compact"
                         variant="underlined"
                         label="Phone Number"
@@ -154,30 +149,27 @@
                         hide-details="auto"
                       />
                     </v-col>
-                    <v-col class="ml-2">
-                      <v-btn
-                        size="x-small"
-                        color="#003366"
-                        variant="flat"
-                        prepend-icon="mdi-content-save"
+                    <v-col cols="3" offset-md="1" class="text-no-wrap">
+                      <AppButton
+                        size="small"
                         type="submit"
+                        :display-block="false"
                         :disabled="!valid.phone"
                         :loading="isProcessing"
                         @click="() => saveField('phone')"
                       >
                         Save
-                      </v-btn>
-                      <v-btn
+                      </AppButton>
+                      <AppButton
                         class="ml-1"
-                        size="x-small"
-                        color="#003366"
-                        variant="outlined"
-                        prepend-icon="mdi-cancel"
+                        size="small"
+                        :display-block="false"
                         :disabled="isProcessing"
+                        :primary="false"
                         @click="editing.phone = false"
                       >
                         Cancel
-                      </v-btn>
+                      </AppButton>
                     </v-col>
                   </v-row>
                 </v-form>
@@ -185,21 +177,19 @@
             </template>
             <template v-else>
               <v-col cols="12" sm="7" xl="8" xxl="9">
-                <v-row no-gutters justify="space-between">
-                  <v-col>
+                <v-row no-gutters>
+                  <v-col cols="12" md="6">
                     <p>{{ loadedModel.phone }}</p>
                   </v-col>
-                  <v-col class="ml-2">
-                    <v-btn
-                      size="x-small"
-                      color="#003366"
-                      variant="outlined"
-                      prepend-icon="mdi-pencil"
+                  <v-col cols="3" offset-md="1">
+                    <AppButton
+                      size="small"
+                      :display-block="false"
                       :disabled="isProcessing"
                       @click="editing.phone = true"
                     >
                       Edit
-                    </v-btn>
+                    </AppButton>
                   </v-col>
                 </v-row>
               </v-col>
@@ -212,11 +202,11 @@
             <template v-if="editing.website">
               <v-col cols="12" sm="7" xl="8" xxl="9">
                 <v-form v-model="valid.website" @submit.prevent>
-                  <v-row no-gutters justify="space-between">
-                    <v-col>
+                  <v-row no-gutters>
+                    <v-col cols="12" md="6">
                       <v-text-field
                         v-model="workingFields.website"
-                        class="micro"
+                        class="less-jitter"
                         density="compact"
                         variant="underlined"
                         label="Website Address"
@@ -225,30 +215,28 @@
                         hide-details="auto"
                       />
                     </v-col>
-                    <v-col class="ml-2">
-                      <v-btn
-                        size="x-small"
+                    <v-col cols="3" offset-md="1" class="text-no-wrap">
+                      <AppButton
+                        size="small"
                         color="#003366"
-                        variant="flat"
-                        prepend-icon="mdi-content-save"
                         type="submit"
+                        :display-block="false"
                         :disabled="!valid.website"
                         :loading="isProcessing"
                         @click="() => saveField('website')"
                       >
                         Save
-                      </v-btn>
-                      <v-btn
+                      </AppButton>
+                      <AppButton
                         class="ml-1"
-                        size="x-small"
-                        color="#003366"
-                        variant="outlined"
-                        prepend-icon="mdi-cancel"
+                        size="small"
+                        :primary="false"
+                        :display-block="false"
                         :disabled="isProcessing"
                         @click="editing.website = false"
                       >
                         Cancel
-                      </v-btn>
+                      </AppButton>
                     </v-col>
                   </v-row>
                 </v-form>
@@ -256,21 +244,20 @@
             </template>
             <template v-else>
               <v-col cols="12" sm="7" xl="8" xxl="9">
-                <v-row no-gutters justify="space-between">
-                  <v-col>
+                <v-row no-gutters>
+                  <v-col cols="12" md="6">
                     <p>{{ loadedModel.website || 'N/A' }}</p>
                   </v-col>
-                  <v-col class="ml-2">
-                    <v-btn
-                      size="x-small"
+                  <v-col cols="3" offset-md="1">
+                    <AppButton
+                      size="small"
                       color="#003366"
-                      variant="outlined"
-                      prepend-icon="mdi-pencil"
+                      :display-block="false"
                       :disabled="isProcessing"
                       @click="editing.website = true"
                     >
                       Edit
-                    </v-btn>
+                    </AppButton>
                   </v-col>
                 </v-row>
               </v-col>
@@ -368,12 +355,14 @@ import { useOrganizationStore } from '@/store/ccof/organization.js';
 import { ORGANIZATION_TYPES } from '@/utils/constants.js';
 import rules from '@/utils/rules.js';
 
-import AppLabel from '@/components/util/AppLabel.vue';
+import AppButton from '@/components/guiComponents/AppButton.vue';
+import AppLabel from '@/components/guiComponents/AppLabel.vue';
 import alertMixin from '@/mixins/alertMixin.js';
 
 export default {
   name: 'ManageOrganization',
   components: {
+    AppButton,
     AppLabel,
   },
   mixins: [alertMixin],
@@ -456,14 +445,14 @@ export default {
 };
 </script>
 <style scoped>
-.micro :deep(input) {
+.less-jitter :deep(input) {
   padding-top: 0;
-  font-size: 0.875rem;
+  font-size: 1rem;
+  letter-spacing: 0.03125em;
   min-height: 0;
 }
 
-.micro :deep(label) {
-  font-size: 0.875rem;
-  top: 0 !important;
+.v-row.v-row--dense {
+  min-height: 38px;
 }
 </style>
